requirements.txt
pyzbar==0.1.9
Pillow==10.0.0
colorama==0.4.6
cairosvg==2.7.1


